requirements.txt
numpy
tensorflow
pillow
tqdm
Pillow
h5py
ipykernel
jupyter
matplotlib
numpy
pandas
tqdm
scipy
sklearn
tensorflow
